requirements.txt
numpy==1.19.1
numpy-stl==2.11.2
python-utils==2.4.0
six==1.15.0
